refactor(data-table-manager): restructure files
