+++ packages/components/data-table-manager/src/column-settings-manager/index.js
@@ -1,0 +1,1 @@
+export { default } from './column-settings-manager';
