+++ packages/components/data-table-manager/src/column-settings-manager/messages.js
