--- packages/components/data-table-manager/src/data-table-settings/column-manager/index.js
@@ -1,1 +1,0 @@
-export { default } from './column-manager';
