+++ packages/components/data-table-manager/src/display-settings-manager/constants.js
