+++ packages/components/data-table-manager/src/display-settings-manager/index.js
@@ -1,0 +1,2 @@
+export { default } from './display-settings-manager';
+export * from './constants';
