+++ packages/components/data-table-manager/src/draggable-tag/draggable-tag.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DragDropContext, Droppable } from 'react-beautiful-dnd';
-import { screen, render } from '../../../../../../../test/test-utils';
+import { screen, render } from '../../../../../test/test-utils';
 import DraggableTag from './draggable-tag';
 
 const createTestProps = (props) => ({
