+++ packages/components/data-table-manager/src/droppable-panel/droppable-panel.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Droppable } from 'react-beautiful-dnd';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
-import DraggableTag from './draggable-tag';
+import DraggableTag from '../draggable-tag';
 import TagContainerEditable from './tag-container-editable';
 import messages from './messages';
