+++ packages/components/data-table-manager/src/droppable-panel/droppable-panel.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DragDropContext } from 'react-beautiful-dnd';
-import { screen, render } from '../../../../../../test/test-utils';
+import { screen, render } from '../../../../../test/test-utils';
 import DroppablePanel from './droppable-panel';
 
 const createTestProps = (props) => ({
