+++ packages/components/data-table-manager/src/droppable-panel/index.js
