+++ packages/components/data-table-manager/src/droppable-panel/tag-container-editable.js
