+++ packages/components/data-table-manager/src/index.js
@@ -1,4 +1,6 @@
 export { default } from './data-table-manager';
-export { useRowSelection, useSorting } from '@commercetools-uikit/hooks';
 export { UPDATE_ACTIONS } from './constants';
 export { default as version } from './version';
+
+// Re-exports for convenience
+export { useRowSelection, useSorting } from '@commercetools-uikit/hooks';
