+++ packages/components/data-table-manager/src/settings-container/index.js
@@ -1,0 +1,1 @@
+export { default } from './settings-container';
